+++ packages/client/src/constants.ts
@@ -30,7 +30,9 @@ export const createConstantEntry = <T>(
       return cachedResults.get(ctx)!
     }
 
-    const pallet = ctx.metadata.pallets.find((p) => p.name === palletName)
+    const pallet = ctx.lookup.metadata.pallets.find(
+      (p) => p.name === palletName,
+    )
     const constant = pallet?.constants.find((c) => c.name === name)!
     const result = ctx.dynamicBuilder
       .buildConstant(palletName, name)