+++ packages/codegen/src/get-new-types.ts
@@ -1,27 +1,26 @@
 import {
   getChecksumBuilder,
-  getLookupFn,
+  MetadataLookup,
 } from "@polkadot-api/metadata-builders"
-import type { V14, V15 } from "@polkadot-api/substrate-bindings"
+import type { V15 } from "@polkadot-api/substrate-bindings"
 import { KnownTypes } from "./known-types"
 import { defaultDeclarations, getTypesBuilder } from "./types-builder"
 
 type ArraVal<T extends Array<any>> = T extends Array<infer V> ? V : unknown
 
 export const getNewTypes = (
-  metadata: V14 | V15,
+  lookup: MetadataLookup,
   knownTypes: KnownTypes,
   getTypeName: (data: ArraVal<V15["lookup"]>) => string | null,
 ) => {
-  const lookup = getLookupFn(metadata.lookup)
-  const checksumBuilder = getChecksumBuilder(metadata, lookup)
+  const { metadata } = lookup
+  const checksumBuilder = getChecksumBuilder(lookup)
   let declarations = defaultDeclarations()
   let typesBuilder = getTypesBuilder(
     declarations,
-    metadata,
+    lookup,
     knownTypes,
     checksumBuilder,
-    lookup,
   )
 
   let ignoredIds = new Set<number>(
@@ -79,10 +78,9 @@ export const getNewTypes = (
   declarations = defaultDeclarations()
   typesBuilder = getTypesBuilder(
     declarations,
-    metadata,
+    lookup,
     wannabes,
     checksumBuilder,
-    lookup,
   )
 
   metadata.lookup.forEach(({ id }) => {